file structure refactoring for DataDomain
